requirements.txt
ai21==1.0.7
discord==2.3.0
python-dotenv~=1.0.0
simplechain==0.0.5.7
annoy==1.17.2
Scrapy~=2.8.0
itemadapter~=0.8.0
colorama~=0.4.6
pydantic==1.10.7
numpy==1.24.3
beautifulsoup4==4.12.2